Enable support for rlimit memlock increase
Virtual machines usually require a lot of memory to be locked to use as a guest RAM. Linux usually restricts the amount of memory that can be locked by the process to 64kb, which obviously won't work, even for a simple 128Mb guest. Enable support for the `rlimit_memlock` option that allows to set the amount of the proccess lockable memory as unlimited when the option is set.